+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './excel2md';
+export * from './deps';